Refactor utilities structure
