+++ packages/core/src/core/node.ts
@@ -1,4 +1,4 @@
-import { generateUUID } from "core/utils";
+import { generateUUID } from "utilities/uuid_generator";
 
 export abstract class Node {
   public readonly uuid = generateUUID();
